refacoring moved common classes to package share, added example for Type Conversion
