--- src/main/scala/dotty/misc/CustomTypeConstructor1.scala
@@ -1,4 +1,0 @@
-package misc
-
-class MyClass[A](val value : A)
-  override def toString = s"MyClass with type ${value.getClass} and value $value"
