--- src/main/scala/dotty/misc/Types.scala
@@ -1,3 +1,0 @@
-package misc.types
-
-type Id[A] = A
