--- src/main/scala/scala2/core/misc/Types.scala
@@ -1,3 +1,0 @@
-package scala2.core.misc
-
-type Id[A] = A
