+++ src/main/scala/share/Types.scala
@@ -1,0 +1,3 @@
+package share
+
+type Id[A] = A
